Fix responsiveness issue and decimal rounding, Attempt narration support
Fixed issue in CSS that caused footer area to appear on mobile, and reworked
how decimals are handled by including code to split and round results
that return as an exponent.

Added title attributes to several buttons in an attempt to offer narration
support (will rework after testing outside of VM).