Fix issue when stringing operations after an error
Fixed an issue where stringing operations after receiving an error would
concat the operator to the error and then return NaN. Now the calculator
should not allow stringing together operations if an error is received.

Also made another update to the checkResults function, which will now round
a decimal to 2 places if it has a repeated string of numbers as its end,
i.e. 1.2300001, otherwise it will return the result as an exponential and
then round that if it has a repeated string of numbers at the end.